feat: 동시성 문제 해결
- 데모 애플리케이션은 동시요청이 많으며, Redis 또한 제한적인 환경으로 가정
- 따라서 비관적/낙관적락보단 애플리케이션 레벨에서 최대 멤버 수 초과 문제를 해결하는 방식을 선택
- 참여자 추가시 그룹의 현재 멤버 수를 1 증가시키는 쿼리를 실행하면서, 최대 멤버 수를 초과하는 경우 0을 반환하도록 구현
- 참여자 퇴장시 그룹 멤버 삭제 쿼리를 실행하면서, 실제로 삭제된 행이 없는 경우(=멤버가 없는 경우) 0을 반환하도록 구현

+++ src/main/kotlin/com/jeonyongho/gls/api/adapter/inbound/rqrs/GroupRqRs.kt
@@ -29,7 +29,7 @@ data class GetMyGroupsRs(
     val groupId: Long,
     val name: String,
     val maxMemberCount: Int,
-    val memberCount: Long,
+    val memberCount: Int,
 ) {
     companion object {
         fun from(result: GetMyGroupsResult) = GetMyGroupsRs(
